Refactor merge widgets
