+++ nbdime-web/src/merge/widget/common.ts
@@ -1,0 +1,18 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+'use strict';
+
+
+// Merge classes:
+export const ONEWAY_LOCAL_CLASS = 'jp-Merge-oneway-local';
+export const ONEWAY_REMOTE_CLASS = 'jp-Merge-oneway-remote';
+export const TWOWAY_ADDITION_CLASS = 'jp-Merge-twoway-addition';
+export const TWOWAY_DELETION_CLASS = 'jp-Merge-twoway-deletion';
+
+const BASE_MERGE_CLASS = 'jp-Merge-base';
+const LOCAL_MERGE_CLASS = 'jp-Merge-local';
+const REMOTE_MERGE_CLASS = 'jp-Merge-remote';
+const MERGED_MERGE_CLASS = 'jp-Merge-merged';
+
+export const MERGE_CLASSES = [BASE_MERGE_CLASS, LOCAL_MERGE_CLASS,
+    REMOTE_MERGE_CLASS, MERGED_MERGE_CLASS];
