+++ nbdime-web/src/merge/widget/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+export * from './cell';
+export * from './notebook';
+export * from './metadata';
